requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
openai==1.54.4
httpx==0.24.1
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
slowapi==0.1.9
python-multipart==0.0.6
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
alembic==1.13.0 